+++ packages/column-views/CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file. The format
 is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this
 project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+- Add mouseover handlers to allow age cursor to be reported
+- Add an `AgeCursor` component
+
 ## [2.0.1] - 2025-05-08
 
 Solve a problem with strict mode